refactor(#16): drop single-impl repository interfaces in :domain
Removes the XxxRepository interfaces in :domain that existed solely as
mocking seams for ViewModel tests. Each repository now ships as a single
concrete class injected via @Inject constructor + @Singleton, removing
the navigation/IDE cost of two files per resource with no runtime payoff.

- Delete DealsRepository / StoresRepository / GamesRepository /
  ReleasesRepository / GiveawaysRepository interface files.
- Rename XxxRepositoryImpl -> XxxRepository (concrete) for all five.
  Constructors are kept @Inject internal constructor so the public class
  is injectable from feature modules without leaking internal DAO types.
- Drop the corresponding @Provides functions from InternalDomainModule;
  Hilt now resolves repositories via @Inject constructor on the class.
- Rename XxxRepositoryImplTest -> XxxRepositoryTest in the :domain
  module and update construction sites accordingly. Test logic is
  unchanged; mockk continues to mock the now-concrete classes.

Feature ViewModels and their tests required no source changes: the
type names match after rename, and existing mockk<XxxRepository>()
calls compile against the concrete class.

Closes #16
