+++ domain/src/main/java/pm/bam/gamedeals/domain/repositories/games/GamesRepository.kt
@@ -1,25 +1,44 @@
 package pm.bam.gamedeals.domain.repositories.games
 
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.onStart
 import kotlinx.serialization.ExperimentalSerializationApi
+import pm.bam.gamedeals.domain.db.dao.GamesDao
 import pm.bam.gamedeals.domain.models.Deal
 import pm.bam.gamedeals.domain.models.Game
 import pm.bam.gamedeals.domain.models.GameDetails
 import pm.bam.gamedeals.domain.models.SearchParameters
+import pm.bam.gamedeals.domain.source.CheapsharkSource
+import javax.inject.Inject
+import javax.inject.Singleton
 
-interface GamesRepository {
+@Singleton
+class GamesRepository @Inject internal constructor(
+    private val gamesDao: GamesDao,
+    private val cheapsharkSource: CheapsharkSource
+) {
 
-    fun observeGames(): Flow<List<Game>>
+    fun observeGames(): Flow<List<Game>> =
+        gamesDao.observeAllGames()
+            .onStart { refreshGames() }
 
-    suspend fun searchGames(query: String): List<Game>
+    suspend fun searchGames(query: String): List<Game> =
+        cheapsharkSource.fetchGames(query)
 
-    @OptIn(ExperimentalSerializationApi::class)
-    suspend fun searchGames(searchParameters: SearchParameters): List<Deal>
+    @ExperimentalSerializationApi
+    suspend fun searchGames(searchParameters: SearchParameters): List<Deal> =
+        cheapsharkSource.fetchDealsForStore(searchParameters)
 
-    suspend fun getReleaseGameId(gameTitle: String): Int?
+    @ExperimentalSerializationApi
+    suspend fun getReleaseGameId(gameTitle: String): Int? =
+        cheapsharkSource.fetchDealsForStore(SearchParameters(title = gameTitle, exact = true))
+            .firstOrNull()
+            ?.gameID
 
-    suspend fun getGameDetails(dealId: Int): GameDetails
+    suspend fun getGameDetails(dealId: Int): GameDetails =
+        cheapsharkSource.fetchGameDetails(dealId.toString())
 
-    suspend fun refreshGames()
-
+    suspend fun refreshGames() =
+        cheapsharkSource.fetchGames("")
+            .let { gamesDao.addGames(*it.toTypedArray()) }
 }
